--- app/src/main/java/com/tao/wnc/NotificationsFragment.java
@@ -1,47 +1,0 @@
-package com.tao.wnc;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import androidx.annotation.NonNull;
-import androidx.databinding.DataBindingUtil;
-import androidx.fragment.app.Fragment;
-
-import com.tao.wnc.databinding.FragmentNotificationsBinding;
-
-/**
- * A simple {@link Fragment} subclass.
- * Use the {@link NotificationsFragment#newInstance} factory method to
- * create an instance of this fragment.
- */
-public class NotificationsFragment extends Fragment {
-
-    public NotificationsFragment() {
-        // Required empty public constructor
-    }
-
-    public static NotificationsFragment newInstance() {
-        return new NotificationsFragment();
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-    }
-
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        FragmentNotificationsBinding binding = DataBindingUtil.inflate(inflater, R.layout.fragment_notifications, container, false);
-        binding.setFragment(this);
-
-        return binding.getRoot();
-    }
-
-    public void onUndoClick(View v) {
-        ((MainActivity) getActivity()).removeAndPop(this);
-    }
-
-}
