+++ app/src/main/java/com/tao/wnc/view/AddPostFragment.java
@@ -1,4 +1,4 @@
-package com.tao.wnc;
+package com.tao.wnc.view;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -9,6 +9,7 @@ import androidx.annotation.NonNull;
 import androidx.databinding.DataBindingUtil;
 import androidx.fragment.app.Fragment;
 
+import com.tao.wnc.R;
 import com.tao.wnc.databinding.FragmentAddPostBinding;
 
 /**
