--- app/src/main/res/drawable/ib_fg.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:drawable="@drawable/ib_fg_after" android:state_pressed="true"/>
-    <item android:drawable="@drawable/ib_fg_after" android:state_focused="true"/>
-    <item android:drawable="@drawable/ib_fg_before"/>
-</selector>
