--- app/src/main/res/drawable/ib_fg_before.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <solid android:color="@color/wncNone"/>
-</shape>
