+++ app/src/main/res/drawable/ic_sentiment_dissatisfied.xml
@@ -1,0 +1,15 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M15.5,9.5m-1.5,0a1.5,1.5 0,1 1,3 0a1.5,1.5 0,1 1,-3 0"
+      android:fillColor="#ffffff"/>
+  <path
+      android:pathData="M8.5,9.5m-1.5,0a1.5,1.5 0,1 1,3 0a1.5,1.5 0,1 1,-3 0"
+      android:fillColor="#ffffff"/>
+  <path
+      android:pathData="M11.99,2C6.47,2 2,6.48 2,12s4.47,10 9.99,10C17.52,22 22,17.52 22,12S17.52,2 11.99,2zM12,20c-4.42,0 -8,-3.58 -8,-8s3.58,-8 8,-8 8,3.58 8,8 -3.58,8 -8,8zM12,16.5c0.73,0 1.39,0.19 1.97,0.53 0.12,-0.14 0.86,-0.98 1.01,-1.14 -0.85,-0.56 -1.87,-0.89 -2.98,-0.89 -1.11,0 -2.13,0.33 -2.99,0.88 0.97,1.09 0.01,0.02 1.01,1.14 0.59,-0.33 1.25,-0.52 1.98,-0.52z"
+      android:fillColor="#ffffff"/>
+</vector>
