+++ app/src/main/res/layout/activity_login.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".LoginActivity">
+        tools:context=".view.LoginActivity">
 
         <FrameLayout
             android:id="@+id/login_container"
