+++ app/src/main/res/layout/fragment_add_post.xml
@@ -6,13 +6,13 @@
     <data>
         <variable
             name="fragment"
-            type="com.tao.wnc.AddPostFragment" />
+            type="com.tao.wnc.view.AddPostFragment" />
     </data>
     
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MainActivity">
+        tools:context=".view.MainActivity">
 
         <EditText
             android:id="@+id/add_post_password"
