+++ app/src/main/res/layout/fragment_login.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="fragment"
-            type="com.tao.wnc.LoginFragment" />
+            type="com.tao.wnc.view.LoginFragment" />
     </data>
 
 
@@ -31,7 +31,7 @@
                 android:layout_height="wrap_content">
 
                 <TextView
-                    android:id="@+id/login_tv_logo"
+                    android:id="@+id/tv_login_logo"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_marginTop="77dp"
@@ -46,7 +46,7 @@
                     app:layout_constraintTop_toTopOf="parent" />
 
                 <androidx.appcompat.widget.AppCompatEditText
-                    android:id="@+id/login_username"
+                    android:id="@+id/edt_login_username"
                     android:layout_width="0dp"
                     android:layout_height="wrap_content"
                     android:layout_marginTop="212dp"
@@ -62,13 +62,13 @@
                     app:backgroundTint="@color/wncLightGray"
                     app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toTopOf="@id/login_tv_logo"
+                    app:layout_constraintTop_toTopOf="@id/tv_login_logo"
                     app:drawableLeftCompat="@drawable/ic_email"
                     app:drawableStartCompat="@drawable/ic_email"
                     android:drawablePadding="5dp"/>
 
                 <androidx.appcompat.widget.AppCompatEditText
-                    android:id="@+id/login_password"
+                    android:id="@+id/edt_login_password"
                     android:layout_width="0dp"
                     android:layout_height="wrap_content"
                     android:layout_marginTop="13dp"
@@ -84,7 +84,7 @@
                     app:backgroundTint="@color/wncLightGray"
                     app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toBottomOf="@+id/login_username"
+                    app:layout_constraintTop_toBottomOf="@+id/edt_login_username"
                     app:drawableLeftCompat="@drawable/ic_lock"
                     app:drawableStartCompat="@drawable/ic_lock"
                     android:drawablePadding="5dp" />
@@ -100,16 +100,15 @@
                     android:text="@string/login_btn_login"
                     android:textSize="18sp"
                     android:textColor="@color/wncWhite"
-
                     app:layout_constraintBottom_toBottomOf="parent"
                     app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toBottomOf="@+id/login_password"
+                    app:layout_constraintTop_toBottomOf="@+id/edt_login_password"
                     app:layout_constraintVertical_bias="0.0" />
 
 
                 <LinearLayout
-                    android:id="@+id/login_ll"
+                    android:id="@+id/ll_login"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_marginTop="8dp"
@@ -142,7 +141,7 @@
                 <LinearLayout
                     android:layout_width="wrap_content"
                     android:layout_height="500dp"
-                    app:layout_constraintTop_toBottomOf="@id/login_ll"
+                    app:layout_constraintTop_toBottomOf="@id/ll_login"
                     tools:ignore="MissingConstraints" />
             </androidx.constraintlayout.widget.ConstraintLayout>
         </ScrollView>
