+++ app/src/main/res/layout/fragment_my_posts.xml
@@ -6,13 +6,13 @@
     <data>
         <variable
             name="fragment"
-            type="com.tao.wnc.MyPostsFragment" />
+            type="com.tao.wnc.view.MyPostsFragment" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MainActivity">
+        tools:context=".view.MainActivity">
 
         <EditText
             android:id="@+id/my_posts_password"
