+++ app/src/main/res/layout/fragment_notifications.xml
@@ -6,13 +6,13 @@
     <data>
         <variable
             name="fragment"
-            type="com.tao.wnc.NotificationsFragment" />
+            type="com.tao.wnc.view.NotificationsFragment" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MainActivity">
+        tools:context=".view.MainActivity">
 
         <EditText
             android:id="@+id/notifications_password"
